V0.1.0 - The beginning
Some work has already been done before I decided to go and put it on GitHub. From hence forward, the versions will change. So far, the project can:

* Create PlayerInteractEvent
* Create PlayerJoinEvent
* Create p.sendMessage() Effect
* Create Checking For Material in Players Hand Trigger
* Create plugin.yml
AND
* Create pom.xml

Essentially, it can create a basic plugin that can check for right/left clicking on blocks/air and send a player a message AND/OR when the player joins the game it will send a message. A few more implements will be coming shortly.