V0.6.0 - Player Health Robustness
Added the ability to set if taking away health can kill or if adding health can max out the players health. Will most likely be defaults when adding this in the GUI later, but will be unselectable if choosen too.

This added an if statement in the Template class to be able to add additional if statements in Java for checking and setting health